+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - allow consuming apps to access handlbars
 - update dependencies
+- allow user config override of commonjs plugin config
 
 ### Changed
 - better error output for css / js build scripts